+++ scim2_cli/__init__.py
@@ -1,8 +1,10 @@
 import json
+import re
 
 import click
 from httpx import Client
 from scim2_client import SCIMClient
+from scim2_models import EnterpriseUser
 from scim2_models import Group
 from scim2_models import User
 from sphinx_click.rst_to_ansi_formatter import make_rst_to_ansi_formatter
@@ -25,9 +27,9 @@ def cli(ctx, url):
     ctx.ensure_object(dict)
     ctx.obj["URL"] = url
     client = Client(base_url=ctx.obj["URL"])
-    ctx.obj["client"] = SCIMClient(client, resource_types=(User, Group))
+    ctx.obj["client"] = SCIMClient(client, resource_types=(User[EnterpriseUser], Group))
     ctx.obj["resource_types"] = {
-        resource_type.__name__.lower(): resource_type
+        re.sub(r"\[.*\]", "", resource_type.__name__.lower()): resource_type
         for resource_type in ctx.obj["client"].resource_types
     }