+++ scim2_cli/delete.py
@@ -1,11 +1,11 @@
-import sys
-
 import click
 from click import ClickException
 from scim2_client import SCIMClientError
 from scim2_models import Message
 from scim2_models import Resource
 from sphinx_click.rst_to_ansi_formatter import make_rst_to_ansi_formatter
+
+from scim2_cli.utils import exception_to_click_error
 
 from .utils import DOC_URL
 from .utils import formatted_payload
@@ -40,13 +40,7 @@ def delete_cli(ctx, resource_type, id, indent):
         response = ctx.obj["client"].delete(resource_model, id, raise_scim_errors=False)
 
     except SCIMClientError as scim_exc:
-        message = str(scim_exc)
-        if sys.version_info >= (3, 11) and hasattr(
-            scim_exc, "__notes__"
-        ):  # pragma: no cover
-            for note in scim_exc.__notes__:
-                message = f"{message}\n{note}"
-        raise ClickException(message) from scim_exc
+        raise exception_to_click_error(scim_exc) from scim_exc
 
     if response:
         payload = (